function-runtime/requirements.txt
flask==3.0.0
google-cloud-storage==2.10.0